refactor: Use kcd-scripts instead of doing everything manually
