+++ .all-contributorsrc
@@ -1,9 +1,15 @@
 {
+  "projectName": "eslint-plugin-jest-dom",
+  "projectOwner": "testing-library",
+  "commit": false,
+  "contributorsPerLine": 7,
+  "imageSize": 100,
+  "repoHost": "https://github.com",
+  "repoType": "github",
+  "skipCi": true,
   "files": [
     "README.md"
   ],
-  "imageSize": 100,
-  "commit": false,
   "contributors": [
     {
       "login": "benmonro",
@@ -48,10 +54,4 @@
       ]
     }
   ],
-  "contributorsPerLine": 7,
-  "projectName": "eslint-plugin-jest-dom",
-  "projectOwner": "testing-library",
-  "repoType": "github",
-  "repoHost": "https://github.com",
-  "skipCi": true
 }
