--- .eslintrc.json
@@ -1,21 +1,0 @@
-{
-    "parserOptions": {
-        "ecmaVersion": 2018
-    },
-    "env": {
-        "commonjs": true,
-        "es6": true,
-        "node": true
-    },
-    "plugins": [
-        "eslint-plugin",
-        "prettier"
-    ],
-    "extends": [
-        "plugin:eslint-plugin/recommended",
-        "prettier"
-    ],
-    "rules": {
-        "no-var": "error"
-    }
-}
