+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,38 @@
+<!--
+Thanks for your interest in the project. I appreciate bugs filed and PRs submitted!
+Please make sure that you are familiar with and follow the Code of Conduct for
+this project (found in the CODE_OF_CONDUCT.md file).
+
+Please fill out this template with all the relevant information so we can
+understand what's going on and fix the issue.
+
+I'll probably ask you to submit the fix (after giving some direction). If you've
+never done that before, that's great! Check this free short video tutorial to
+learn how: http://kcd.im/pull-request
+-->
+
+- `eslint-plugin-jest-dom` version:
+- `node` version:
+- `npm` version:
+
+Relevant code or config
+
+```javascript
+```
+
+What you did:
+
+What happened:
+
+<!-- Please provide the full error message/screenshots/anything -->
+
+Reproduction repository:
+
+<!--
+If possible, please create a repository that reproduces the issue with the
+minimal amount of code possible.
+-->
+
+Problem description:
+
+Suggested solution:
