+++ .gitignore
@@ -1,5 +1,9 @@
 node_modules
 coverage
-yarn.lock
+dist
+.DS_Store
+
+# these cause more harm than good
+# when working with contributors
 package-lock.json
-.vscode
+yarn.lock
