+++ __tests__/fixtures/createBannedAttributeTestCases.js
