+++ jest.config.js
@@ -1,14 +1,13 @@
+const jestConfig = require('kcd-scripts/jest')
+
 module.exports = {
-  testMatch: ["**/tests/**/*.js"],
-  collectCoverage: true,
+  ...jestConfig,
   coverageThreshold: {
     global: {
       branches: 96.55,
       functions: 100,
       lines: 98.97,
-      statements: 0
-    }
+      statements: 0,
+    },
   },
-  testPathIgnorePatterns: ["<rootDir>/tests/fixtures/"],
-  collectCoverageFrom: ["lib/**/*.js", "!**/node_modules/**"]
 };
