+++ other/USERS.md
@@ -1,0 +1,12 @@
+# Users
+
+If you or your company uses this project, add your name to this list! Eventually
+we may have a website to showcase these (wanna build it!?)
+
+> No users have been added yet!
+
+<!--
+This file should just be a bulleted list like this:
+
+- [Company/Project/Person](https://example.com) uses it in [some app](https://example.com)
+-->
