+++ src/createBannedAttributeRule.js
