+++ src/index.js
@@ -14,7 +14,7 @@ let requireIndex = require("requireindex");
 // Plugin Definition
 //------------------------------------------------------------------------------
 
-// import all rules in lib/rules
+// import all rules in src/rules
 module.exports.rules = requireIndex(__dirname + "/rules");
 
 module.exports.generateRecommendedConfig = rules => {
