+++ src/rules/prefer-checked.js
