+++ src/rules/prefer-empty.js
