+++ src/rules/prefer-focus.js
