+++ src/rules/prefer-to-have-text-content.js
